backend/farmconnect_backend/backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# AI and ML
openai==1.54.3
google-generativeai==0.8.3
langchain==0.3.7
langchain-openai==0.2.8
langchain-google-genai==2.0.4

# HTTP Requests
httpx==0.27.2
requests==2.32.3
aiohttp==3.11.7

# Data Validation
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# CORS and Middleware
# python-cors==1.0.0

# Caching
redis==5.2.0
hiredis==3.0.0

# Rate Limiting
slowapi==0.1.9

# Logging and Monitoring
loguru==0.7.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Code Quality
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Utils
python-dateutil==2.9.0.post0
pytz==2024.2

greenlet==3.1.1